Add proxy option for forms
post_app_proxy can now be set as an easy way to tell the front-end that post requests should be made via a proxy. The proxy path is local to the front-end and validation remains the same. This is useful for a registration request where the final post data should be sent via a proxy so as to not expose the refresh token