Mapping custom component entities
This introduces automated discriminator mapping for components which means that any new component created at application level will be added to the discriminator map. It is quite expensive (computation wise) and needs to be improved so that the default mapping is respected and it can be merged with additional components from the config. Alternatively, there are possibilities to cache the entities that extend the AbstractComponent or implement ComponentInterface and possibly add annotation so that the name in the discriminator column can be customised. This needed to be released as a matter of urgency though so has been implemented in the same way that Doctrine automatically generates the discriminator map but using snake_case instead and forcing the mapping to be regenreated.