Filtering via querystring for Collection entity
This update allows the filter querystring for a collection to be passed via the Collection entity. The Collection entity is a component which includes a collection of another entity (whether this is a DynamicContent entity or any other component).
If you configure a filter for the entity in the collection, you can now still use that filter against the 'Collection' entity and this will simply pass it through to the collection provider.