Data Transformers and Security Updates
Security - DenyAccessListener for new RestrictedResourceInterface implementation. This allows a resource to be restricted to specific roles. This is implemented by default on the AbstractContent class (and descendants) Therefore we also check DynamicContent, Routes, ComponentLocations and any components extending AbstractComponent to ensure they are either on a public page, or if they are only located on restricted pages, those restricted pages roles are enforced.
DataTransformers - Work carried out to improve DataTransformers (formerly DataModifiers). Had implemented API Platform DTO methods, but this resulted in loss of IRI. Similar implementation and improved ApiNormalizer using best practices.