Restricted Resource improvements
If a component is a restricted resource, when a page is loading (i.e. during normalization so if the resource is anywhere in the tree) it will be stripped out. This is a different behaviour from a restricted resource that is a page which will throw an access denied excexception.
The security roles in the restricted resources can now use an exclamation to negate that role from viewing the component. E.g. ['ROLE_ADMIN', '!ROLE_USER'] - ROLE_ADMIN will view the component for editing, if the user just has ROLE_USER it will not display. But it would still display to a non-logged in user.
A RestrictedResourceTrait has been added to make it easier to implement