Get sort value using symfony property accessor
We do not need to get a sort value only if it uses the SortableTrait. So long as we can access 'sort' and the property accessor is more robust at finding properties in objects or arrays. It may be we want to make sure a collection is only the same resource... but it may not matter and this could be more flexible.