AdminContextBuilder change to UserContextBuilder
To prevent the possibility of a security flaw with groups just defining the security role (and additional unnecessary serialization) this commit changes the serialization context builder to specifically focus on the groups for the user class and the user class now defines groups specific to the class as well. There are still additional field permissions for users with ROLE_SUPER_ADMIN. We may want to make the supported role configurable in future.

Also see issue https://github.com/silverbackis/ApiComponentBundle/issues/34

Context Builder unit tests added